s2i doesn't support root commands
Give up trying to install anything with yum, etc. It's against the philosophy of the `source-to-image` tool (the whole point is for it to be fast and "safe" for developers to use).

Instead, this kind of logic I have been aiming for, should be placed in the source-to-image **base** image (which is the basis of the builds for the dev teams). See here about how to create a base image:
https://blog.openshift.com/create-s2i-builder-image/

In addition, when digging deeper, I realized that the PHP base image I was using already includes the mysql binding: https://github.com/sclorg/s2i-php-container/blob/master/7.1/Dockerfile